+++ mojo/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "1.2.39"
+__version__ = "1.2.40"
 
 from mojo.helpers.response import JsonResponse